moved disk ICs into new file disk_ICs.cpp
